Move debugger registration to separate file.
